requirements.txt
aio-pika==9.5.5
aiokafka==0.12.0
aiormq==6.8.1
alembic==1.15.1
amqp==5.3.1
annotated-types==0.7.0
anyio==4.5.2
async-timeout==5.0.1
asyncpg==0.30.0
billiard==4.2.1
black==25.1.0
celery==5.5.2
certifi==2025.1.31
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
dnspython==2.6.1
ecdsa==0.19.1
email_validator==2.2.0
exceptiongroup==1.2.2
factory_boy==3.3.3
Faker==37.1.0
fastapi==0.115.11
fastapi-cli==0.0.7
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
kombu==5.5.3
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.4.3
mypy_extensions==1.1.0
orjson==3.10.15
packaging==25.0
pamqp==3.3.0
pathspec==0.12.1
platformdirs==4.3.8
pluggy==1.5.0
prompt_toolkit==3.0.51
propcache==0.3.1
psycopg2==2.9.10
pyasn1==0.4.8
pydantic==2.10.6
pydantic-extra-types==2.10.2
pydantic-settings==2.8.1
pydantic_core==2.27.2
Pygments==2.19.1
pytest==8.3.5
pytest-asyncio==0.26.0
pytest-factoryboy==2.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
PyYAML==6.0.2
redis==5.2.1
rich==13.9.4
rich-toolkit==0.13.2
rsa==4.9
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.39
starlette==0.44.0
typer==0.15.2
typing_extensions==4.12.2
tzdata==2025.2
ujson==5.10.0
uvicorn==0.33.0
uvloop==0.21.0
vine==5.1.0
watchfiles==0.24.0
wcwidth==0.2.13
websockets==13.1
yarl==1.20.0
